Added Asteroids,Asteroid Animations and random asteroid generator.
There is still a bug with the Random asteroid generator as it generates asteroids off screen. Since this is tied to a random number generator, What we can do to achieve the same output is to have "zones" on the screen which will create asteroids. These "Zones" will have variable intervals to instantiate an asteroid. Further testing required.